pyproject.toml
[tool.poetry]
name = "coursework-3"
version = "0.1.0"
description = "coursework_3"
authors = ["Artur Koltyrin <koltyrinartur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
